tools/local_loader/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
azure-storage-blob==12.16.0